+++ src/main/java/Application.java
@@ -5,8 +5,8 @@ import db.dao.GameDao;
 import db.jdbc.ConnectionManager;
 import db.jdbc.DatabaseMigrator;
 import db.jdbc.FlywayDatabaseMigrator;
-import db.jdbc.JdbcBoardPieceDao;
-import db.jdbc.JdbcGameDao;
+import db.dao.JdbcBoardPieceDao;
+import db.dao.JdbcGameDao;
 import db.jdbc.ProductionConnectionManager;
 import db.repository.JanggiGameRepository;
 import db.repository.JdbcJanggiGameRepository;
