+++ src/test/java/db/repository/JdbcJanggiGameRepositoryTest.java
@@ -12,8 +12,8 @@ import db.dao.GameDao;
 import db.jdbc.ConnectionManager;
 import db.jdbc.DatabaseMigrator;
 import db.jdbc.FlywayDatabaseMigrator;
-import db.jdbc.JdbcBoardPieceDao;
-import db.jdbc.JdbcGameDao;
+import db.dao.JdbcBoardPieceDao;
+import db.dao.JdbcGameDao;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
